Make triangle geometries more modulable.
