+++ src/webots/Makefile
@@ -438,6 +438,7 @@ QT_SOURCES += \
   WbTranslateRotateManipulator.cpp \
   WbTreeItem.cpp \
   WbTreeView.cpp \
+	WbTriangleMeshGeometry.cpp \
   WbUndoStack.cpp \
   WbValue.cpp \
   WbValueEditor.cpp \
